# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.EnterpriseLoginRequest:
    properties:
      device_fingerprint:
        $ref: '#/definitions/services.DeviceFingerprint'
      device_name:
        type: string
      device_type:
        description: mobile/desktop/tablet
        type: string
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  services.DeviceFingerprint:
    properties:
      ip_address:
        type: string
      language:
        type: string
      platform:
        type: string
      screen:
        additionalProperties:
          type: integer
        type: object
      timezone:
        type: string
      user_agent:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@split-go.com
    name: Split Go API Support
    url: http://www.split-go.com/support
  description: 分帳系統 API 文檔 - 支援群組管理、交易記錄、結算功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Split Go API
  version: "1.0"
paths:
  /auth/device-refresh:
    post:
      consumes:
      - application/json
      description: 使用設備指紋進行刷新，適用於記住登入狀態的場景
      parameters:
      - description: 設備刷新請求
        in: body
        name: request
        required: true
        schema:
          properties:
            device_fingerprint:
              properties:
                ip_address:
                  type: string
                language:
                  type: string
                platform:
                  type: string
                screen_resolution:
                  type: string
                timezone:
                  type: string
                user_agent:
                  type: string
              type: object
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功，返回新的令牌對
          schema:
            properties:
              data:
                properties:
                  access_token:
                    type: string
                  device_token:
                    type: string
                  expires_in:
                    type: integer
                  refresh_token:
                    type: string
                  session:
                    properties:
                      device_name:
                        type: string
                      device_type:
                        type: string
                      expires_at:
                        type: string
                      id:
                        type: string
                      last_activity:
                        type: string
                    type: object
                  user:
                    properties:
                      avatar:
                        type: string
                      email:
                        type: string
                      id:
                        type: integer
                      name:
                        type: string
                      username:
                        type: string
                    type: object
                type: object
              error:
                type: boolean
              message:
                type: string
            type: object
        "400":
          description: 請求格式錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: 設備認證失敗
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      summary: 設備級刷新令牌
      tags:
      - 認證
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用 email 和密碼登入系統，支援設備指紋識別
      parameters:
      - description: 登入資料
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.EnterpriseLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登入成功，返回完整令牌信息
          schema:
            properties:
              data:
                properties:
                  access_token:
                    type: string
                  device_token:
                    type: string
                  expires_in:
                    type: integer
                  refresh_token:
                    type: string
                  session:
                    properties:
                      device_name:
                        type: string
                      device_type:
                        type: string
                      expires_at:
                        type: string
                      id:
                        type: string
                      last_activity:
                        type: string
                    type: object
                  user:
                    properties:
                      avatar:
                        type: string
                      created_at:
                        type: string
                      email:
                        type: string
                      id:
                        type: integer
                      name:
                        type: string
                      updated_at:
                        type: string
                      username:
                        type: string
                    type: object
                type: object
              error:
                type: boolean
              message:
                type: string
            type: object
        "400":
          description: 請求格式錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: 認證失敗
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      summary: 用戶登入
      tags:
      - 認證
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 登出當前用戶，撤銷當前會話的所有令牌
      parameters:
      - description: 登出請求
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "400":
          description: 請求格式錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 用戶登出
      tags:
      - 認證
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用 refresh token 獲取新的 access token 和 refresh token
      parameters:
      - description: 刷新令牌請求
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功，返回新的令牌對
          schema:
            properties:
              data:
                properties:
                  access_token:
                    type: string
                  device_token:
                    type: string
                  expires_in:
                    type: integer
                  refresh_token:
                    type: string
                  session:
                    properties:
                      device_name:
                        type: string
                      device_type:
                        type: string
                      expires_at:
                        type: string
                      id:
                        type: string
                      last_activity:
                        type: string
                    type: object
                  user:
                    properties:
                      avatar:
                        type: string
                      email:
                        type: string
                      id:
                        type: integer
                      name:
                        type: string
                      username:
                        type: string
                    type: object
                type: object
              error:
                type: boolean
              message:
                type: string
            type: object
        "400":
          description: 請求格式錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: refresh token 無效或已過期
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      summary: 刷新訪問令牌
      tags:
      - 認證
  /auth/register:
    post:
      consumes:
      - application/json
      description: 創建新的用戶帳號
      parameters:
      - description: 註冊資料
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              type: string
            name:
              type: string
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 註冊成功
          schema:
            properties:
              data:
                properties:
                  avatar:
                    type: string
                  created_at:
                    type: string
                  email:
                    type: string
                  id:
                    type: integer
                  name:
                    type: string
                  updated_at:
                    type: string
                  username:
                    type: string
                type: object
              error:
                type: boolean
              message:
                type: string
            type: object
        "400":
          description: 請求格式錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "409":
          description: 用戶已存在
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      summary: 用戶註冊
      tags:
      - 認證
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 在 header 中加入 "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"